requirements.txt
django-compat>=1.0.15
apscheduler>=3.7.0